refactor(message): 패키지 이름 error로 변경
